refactor: 방 내 채팅 따로 관리
