+++ src/main/java/com/curioussong/alsongdalsong/game/controller/GameController.java
@@ -3,6 +3,7 @@ package com.curioussong.alsongdalsong.game.controller;
 import com.curioussong.alsongdalsong.game.dto.chat.ChatRequestDTO;
 import com.curioussong.alsongdalsong.game.dto.start.StartRequest;
 import com.curioussong.alsongdalsong.game.dto.userinfo.ReadyRequest;
+import com.curioussong.alsongdalsong.game.service.ChatHandler;
 import com.curioussong.alsongdalsong.game.service.GameService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,6 +18,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class GameController {
 
     private final GameService gameService;
+    private final ChatHandler chatHandler;
 
     @MessageMapping("/channel/{channelId}/room/{roomId}/start")
     @SendTo("/topic/channel/{channelId}/room/{roomId}")
@@ -34,6 +36,6 @@ public class GameController {
 
     @MessageMapping("/channel/{channelId}/room/{roomId}")
     public void handleRoomChat(ChatRequestDTO chatRequestDTO, @DestinationVariable Long channelId, @DestinationVariable String roomId) {
-        gameService.roomChatMessage(chatRequestDTO, channelId, roomId);
+        chatHandler.handleChat(chatRequestDTO, channelId, roomId);
     }
 }
